- issue 2138: added gamma / poisson as allowed expressions for the expression widget, and added SBML import / export similar to what we do for normal / uniform

+++ copasi/function/CEvaluationLexer.h
@@ -152,8 +152,8 @@ public:
   virtual int yywrap() {return 1;}
 
 protected:
-  virtual int LexerInput(char* buf, int max_size);
-  virtual void LexerOutput(const char* buf, int size);
+  virtual size_t LexerInput(char* buf, size_t max_size);
+  virtual void LexerOutput(const char* buf, size_t size);
   virtual void LexerError(const char* msg);
 
   void yyunput(int c, char* buf_ptr);
